% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amiga-package.R
\docType{package}
\name{amiga-package}
\alias{amiga}
\alias{amiga-package}
\title{amiga: Check for leaks in your deSolve function}
\description{
Oh no did you write your equations right in your deSolve function? Check for leaks!
}
\author{
\strong{Maintainer}: Damie Pak \email{damiedpak@gmail.com} (\href{https://orcid.org/0000-0002-8335-110X}{ORCID})

}
\keyword{internal}
