% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_leaks_fast.R
\name{check_leaks_fast}
\alias{check_leaks_fast}
\title{Check leaks in your system of ODEs fast}
\usage{
check_leaks_fast(deSolve_function)
}
\arguments{
\item{deSolve_function}{}
}
\value{
A message saying your equation is fine or not
}
\description{
If you have a system of equations that you know should
add up to 0: this is the simplest function for it. If there is a leak,
will not tell you where. Also not useless if the total population
is not constant.
}
\examples{
check_leaks_fast(SIR_model)
}
