% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adjacency_matrix.R
\name{create_adjacency_matrix}
\alias{create_adjacency_matrix}
\title{Creates an adjacency matrix of the system of equations}
\usage{
create_adjacency_matrix(yacas_list)
}
\arguments{
\item{yacas_list}{List with the first element being the state variables and
the second element being the systems of equations in ryacas form.}
}
\value{
A matrix object
}
\description{
The adjacency matrix shows where the state variables are across
the compartmental model
}
\examples{
create_adjacency_matrix(SIR_list)
}
