% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_system_equations.R
\name{extract_system_equations}
\alias{extract_system_equations}
\title{Extract the system equations from a deSolve function}
\usage{
extract_system_equations(deSolve_function)
}
\arguments{
\item{deSolve_function}{The desolve function}
}
\value{
A list element with the first element being the state variables
and the second element being a list of the equations.
}
\description{
By feeding in a deSolve function, you can then retrieve the state-variables
of interest as well as the corresponding equations related to their derivative.
Note: Because this function using \code{{ryacas}}, it automatically assumes that I
is a complex number. This is very annoying if you're doing SIR, so I automatically
change I to little i.
}
\examples{
extract_system_equations(SIRmod)
}
