% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_adjacency_matrix.R
\name{graph_adjacency_matrix}
\alias{graph_adjacency_matrix}
\title{Plot the adjacency matrix}
\usage{
graph_adjacency_matrix(adjacency_matrix)
}
\arguments{
\item{adjacency_matrix}{}
}
\value{
ggplot2 object that shows the states as rectangles AND how they
are connected or not.
}
\description{
Primitive plotting function as of right now on what your system of equations
look like right now
}
\examples{
graph_adjacency_matrix(adj_mat)
}
